boards/esp32c6: Rename esp32c6-devkit as esp32c6-devkitc
