--- boards/risc-v/esp32c6/esp32c6-devkit/Kconfig
@@ -1,8 +1,0 @@
-#
-# For a description of the syntax of this configuration file,
-# see the file kconfig-language.txt in the NuttX tools repository.
-#
-
-if ARCH_BOARD_ESP32C6_DEVKIT
-
-endif # ARCH_BOARD_ESP32C6_DEVKIT
