--- boards/risc-v/esp32c6/esp32c6-devkit/configs/spiflash/defconfig
@@ -1,55 +1,0 @@
-#
-# This file is autogenerated: PLEASE DO NOT EDIT IT.
-#
-# You can use "make menuconfig" to make any modifications to the installed .config file.
-# You can then do "make savedefconfig" to generate a new defconfig file that includes your
-# modifications.
-#
-# CONFIG_NSH_ARGCAT is not set
-# CONFIG_NSH_CMDOPT_HEXDUMP is not set
-CONFIG_ARCH="risc-v"
-CONFIG_ARCH_BOARD="esp32c6-devkit"
-CONFIG_ARCH_BOARD_COMMON=y
-CONFIG_ARCH_BOARD_ESP32C6_DEVKIT=y
-CONFIG_ARCH_CHIP="esp32c6"
-CONFIG_ARCH_CHIP_ESP32C6=y
-CONFIG_ARCH_INTERRUPTSTACK=2048
-CONFIG_ARCH_RISCV=y
-CONFIG_ARCH_STACKDUMP=y
-CONFIG_BOARDCTL_RESET=y
-CONFIG_BOARD_LOOPSPERMSEC=15000
-CONFIG_BUILTIN=y
-CONFIG_DEV_ZERO=y
-CONFIG_ESPRESSIF_ESP32C6=y
-CONFIG_ESPRESSIF_SPIFLASH=y
-CONFIG_ESPRESSIF_SPIFLASH_SMARTFS=y
-CONFIG_ESPRESSIF_STORAGE_MTD_OFFSET=0x110000
-CONFIG_ESPRESSIF_STORAGE_MTD_SIZE=0xf0000
-CONFIG_FS_PROCFS=y
-CONFIG_IDLETHREAD_STACKSIZE=2048
-CONFIG_INIT_ENTRYPOINT="nsh_main"
-CONFIG_INTELHEX_BINARY=y
-CONFIG_LIBC_PERROR_STDOUT=y
-CONFIG_LIBC_STRERROR=y
-CONFIG_NAME_MAX=48
-CONFIG_NFILE_DESCRIPTORS_PER_BLOCK=6
-CONFIG_NSH_ARCHINIT=y
-CONFIG_NSH_BUILTIN_APPS=y
-CONFIG_NSH_DISABLE_LOSMART=y
-CONFIG_NSH_FILEIOSIZE=512
-CONFIG_NSH_READLINE=y
-CONFIG_NSH_STRERROR=y
-CONFIG_PREALLOC_TIMERS=0
-CONFIG_RR_INTERVAL=200
-CONFIG_SCHED_BACKTRACE=y
-CONFIG_SCHED_WAITPID=y
-CONFIG_SMARTFS_MAXNAMLEN=48
-CONFIG_START_DAY=29
-CONFIG_START_MONTH=11
-CONFIG_START_YEAR=2019
-CONFIG_SYSTEM_DUMPSTACK=y
-CONFIG_SYSTEM_FLASH_ERASEALL=y
-CONFIG_SYSTEM_NSH=y
-CONFIG_TESTING_FSTEST=y
-CONFIG_TESTING_FSTEST_MOUNTPT="/mnt"
-CONFIG_UART0_SERIAL_CONSOLE=y
