--- boards/risc-v/esp32c6/esp32c6-devkit/src/Make.defs
@@ -1,43 +1,0 @@
-#############################################################################
-# boards/risc-v/esp32c6/esp32c6-devkit/src/Make.defs
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.  The
-# ASF licenses this file to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance with the
-# License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
-# License for the specific language governing permissions and limitations
-# under the License.
-#
-#############################################################################
-
-include $(TOPDIR)/Make.defs
-
-CSRCS = esp32c6_boot.c esp32c6_bringup.c
-
-ifeq ($(CONFIG_BOARDCTL),y)
-  CSRCS += esp32c6_appinit.c
-
-  ifeq ($(CONFIG_BOARDCTL_RESET),y)
-    CSRCS += esp32c6_reset.c
-  endif
-endif
-
-ifeq ($(CONFIG_DEV_GPIO),y)
-  CSRCS += esp32c6_gpio.c
-endif
-
-ifeq ($(CONFIG_ARCH_BUTTONS),y)
-  CSRCS += esp32c6_buttons.c
-endif
-
-DEPPATH += --dep-path board
-VPATH += :board
-CFLAGS += ${INCDIR_PREFIX}$(TOPDIR)$(DELIM)arch$(DELIM)$(CONFIG_ARCH)$(DELIM)src$(DELIM)board$(DELIM)board
