+++ boards/risc-v/esp32c6/esp32c6-devkitc/Kconfig
@@ -1,0 +1,8 @@
+#
+# For a description of the syntax of this configuration file,
+# see the file kconfig-language.txt in the NuttX tools repository.
+#
+
+if ARCH_BOARD_ESP32C6_DEVKITC
+
+endif # ARCH_BOARD_ESP32C6_DEVKITC
